1) implementing the 256 queues which will be primed (init) in prime_queues function 2) Added the TIMESLICE constants to calculate from 10 to 200 the slices for each queue.
The idea is that each queue will count with a list_head to track their tasks. We neeed
to implement the functions which will handle the queues switching etc...

Also, the need to define a MACRO to calculate the priority of each queue. By using 256
queues, this value is given by (38*(index+1)+472)/51 - i will add this soon